Part4/Chapter02-01-copilot/requirements.txt
streamlit==1.45.1
dotenv==0.9.9
google-genai==1.13